+++ scripts/generate-starter-kits.js
@@ -40,20 +40,3 @@ async function getKitMetadata(kitDir) {
   console.warn(errorMessage)
   throw new Error(errorMessage)
 }
-
-function getRepoRootPath() {
-  return path.resolve(__dirname, '../');
-}
-
-async function getKitDirs() {
-  const startersPath = path.resolve(getRepoRootPath(), 'starters');
-  const dirItems = await fs.readdir(startersPath);
-  const kitDirs = [];
-  for (const item of dirItems) {
-    const stats = await fs.stat(startersPath + '/' + item);
-    if (stats.isDirectory() && item.charAt(0) !== '.') {
-      kitDirs.push(item);
-    }
-  }
-  return kitDirs;
-}
